refactor: move property reflections into reflections namespace
