+++ src/Reflections/ReflectionTypeContainer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace NunoMaduro\Larastan\Extensions\Properties;
+namespace NunoMaduro\Larastan\Reflections;
 
 use ReflectionNamedType;
